+++ packages/orq-rc/src/models/operations/listagents.ts
@@ -1076,7 +1076,7 @@ export const ListAgentsKnowledgeBases$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  id: z.string().default("01K7KE87TMDRQ6FX4K1TVJK47Q"),
+  id: z.string().default("01K7KP68Y3DDBAJX9ERTP57ZSZ"),
   knowledge_id: z.string(),
   configuration: z.union([
     z.lazy(() =>
@@ -1107,7 +1107,7 @@ export const ListAgentsKnowledgeBases$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   ListAgentsKnowledgeBases
 > = z.object({
-  id: z.string().default("01K7KE87TMDRQ6FX4K1TVJK47Q"),
+  id: z.string().default("01K7KP68Y3DDBAJX9ERTP57ZSZ"),
   knowledgeId: z.string(),
   configuration: z.union([
     z.lazy(() =>