+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -298,7 +298,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-10-15T08:02:04.170Z",
+    "2025-10-15T10:20:49.148Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -338,7 +338,7 @@ export const UpdateDatasetResponseBody$outboundSchema: z.ZodType<
   createdById: z.string().optional(),
   updatedById: z.string().optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-10-15T08:02:04.170Z"))
+  updated: z.date().default(() => new Date("2025-10-15T10:20:49.148Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {