+++ packages/orq-rc/RELEASES.md
@@ -4269,3 +4269,13 @@ Based on:
 - [typescript v3.14.0-rc.27] packages/orq-rc
 ### Releases
 - [NPM v3.14.0-rc.27] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.27 - packages/orq-rc
+
+## 2025-10-15 10:22:37
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.636.3 (2.723.11) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.14.0-rc.28] packages/orq-rc
+### Releases
+- [NPM v3.14.0-rc.28] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.28 - packages/orq-rc